+++ CHANGES.md
@@ -10,6 +10,8 @@
 - Add `magik-typed-lint` to check magik typing from CLI/during CI.
 - Fix not indexing `module.def` files.
 - Rename SwChar16VectorEvaluateInvocationCheck to SwChar16VectorEvaluateInvocationTypedCheck.
+- Make tab width configurable for LineLengthCheck.
+- Extend VariableNamingCheck to also check for a maximum of chars and make minLength/maxLength configurable.
 - Add NestingDepthCheck to test if a method/procedure/if-statement/loop-statement does not exceed the maximum nesting depth.
 - Several fixes.